Added API key and user's IP to the Google Books API calls.
According to the documentation with the GoogleBooks gem:
"The current Google API for Books generally requires that the requesting
server's IP address be geolocatable. This can cause responses to be
restricted when requested from distributed/clustered servers such as Heroku."

So, the user's IP address is now passed through with the API calls and an
API key is used in the hope that it allows more calls to be made.